+++ src/Command/GenerateCommand.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace DoclerLabs\ApiClientGenerator\Command;
 
@@ -13,6 +15,7 @@ use DoclerLabs\ApiClientGenerator\Output\MetaFilePrinter;
 use DoclerLabs\ApiClientGenerator\Output\Php\PhpFileCollection;
 use DoclerLabs\ApiClientGenerator\Output\PhpFilePrinter;
 use DoclerLabs\ApiClientGenerator\Output\StaticPhpFileCopier;
+use DoclerLabs\ApiClientGenerator\Output\WarningFormatter;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -23,16 +26,17 @@ use Symfony\Component\Finder\Finder;
 
 class GenerateCommand extends Command
 {
-    private Configuration        $configuration;
-    private CodeGeneratorFacade  $codeGenerator;
-    private FileReader           $fileReader;
-    private Parser               $parser;
-    private PhpFilePrinter       $phpPrinter;
-    private MetaTemplateFacade   $metaTemplate;
-    private MetaFilePrinter      $templatePrinter;
-    private Finder               $fileFinder;
-    private StaticPhpFileCopier  $staticPhpPrinter;
-    private Filesystem           $filesystem;
+    private Configuration       $configuration;
+    private CodeGeneratorFacade $codeGenerator;
+    private FileReader          $fileReader;
+    private Parser              $parser;
+    private PhpFilePrinter      $phpPrinter;
+    private MetaTemplateFacade  $metaTemplate;
+    private MetaFilePrinter     $templatePrinter;
+    private Finder              $fileFinder;
+    private StaticPhpFileCopier $staticPhpPrinter;
+    private Filesystem          $filesystem;
+    private WarningFormatter    $warningFormatter;
 
     public function __construct(
         Configuration $configuration,
@@ -44,7 +48,8 @@ class GenerateCommand extends Command
         MetaFilePrinter $templatePrinter,
         Finder $fileFinder,
         StaticPhpFileCopier $staticPhpCopier,
-        Filesystem $filesystem
+        Filesystem $filesystem,
+        WarningFormatter $warningFormatter
     ) {
         parent::__construct();
         $this->configuration    = $configuration;
@@ -57,6 +62,7 @@ class GenerateCommand extends Command
         $this->fileFinder       = $fileFinder;
         $this->staticPhpPrinter = $staticPhpCopier;
         $this->filesystem       = $filesystem;
+        $this->warningFormatter = $warningFormatter;
     }
 
     public function configure(): void
@@ -70,6 +76,7 @@ class GenerateCommand extends Command
 
     public function execute(InputInterface $input, OutputInterface $output): int
     {
+        $this->initWarningPrinting($input);
         $specificationFilePath = $this->configuration->getSpecificationFilePath();
 
         $specification = $this->parser->parse(
@@ -161,17 +168,31 @@ class GenerateCommand extends Command
 
     private function copySpecification(StyleInterface $ss): void
     {
-        $ss->text(sprintf('Copy specification file to %s.', $this->configuration->getOutputDirectory()));
-
         $destinationPath = sprintf(
             '%s/doc/%s',
             $this->configuration->getOutputDirectory(),
             basename($this->configuration->getSpecificationFilePath())
         );
 
+        $ss->text(sprintf('Copy specification file to %s.', $destinationPath));
+
         $this->filesystem->copy(
             $this->configuration->getSpecificationFilePath(),
             $destinationPath
         );
+    }
+
+    private function initWarningPrinting(InputInterface $input): void
+    {
+        if ($input->getOption('quiet')) {
+            set_error_handler(
+                static function (): bool {
+                    return true;
+                },
+                E_USER_WARNING
+            );
+        } else {
+            set_error_handler($this->warningFormatter, E_USER_WARNING);
+        }
     }
 }