+++ src/Generator/ClientGenerator.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace DoclerLabs\ApiClientGenerator\Generator;
 
@@ -11,8 +13,8 @@ use DoclerLabs\ApiClientGenerator\Naming\SchemaMapperNaming;
 use DoclerLabs\ApiClientGenerator\Output\Copy\Request\Mapper\RequestMapperInterface;
 use DoclerLabs\ApiClientGenerator\Output\Copy\Request\RequestInterface;
 use DoclerLabs\ApiClientGenerator\Output\Copy\Response\ResponseHandler;
+use DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType\ContentTypeSerializerInterface;
 use DoclerLabs\ApiClientGenerator\Output\Php\PhpFileCollection;
-use phpDocumentor\Reflection\Types\This;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Property;
 use Psr\Container\ContainerInterface;
@@ -55,13 +57,15 @@ class ClientGenerator extends GeneratorAbstract
             'map',
             $this->builder->args([$requestVariable])
         );
+
         $clientCall = $this->builder->methodCall(
             $this->builder->localPropertyFetch('client'),
             'sendRequest',
             [$mapMethodCall]
         );
 
-        return $this->builder->method('sendRequest')
+        return $this->builder
+            ->method('sendRequest')
             ->makePublic()
             ->addParam($methodParameter)
             ->addStmt($this->builder->return($clientCall))
@@ -83,7 +87,7 @@ class ClientGenerator extends GeneratorAbstract
             )
         );
 
-        $requestVar = $this->builder->var('request');
+        $requestVar  = $this->builder->var('request');
         $methodParam = $this->builder
             ->param('request')
             ->setType($requestClassName)
@@ -98,7 +102,8 @@ class ClientGenerator extends GeneratorAbstract
 
         $responseBody = $operation->getSuccessfulResponse()->getBody();
         if ($responseBody === null) {
-            return $this->builder->method($operation->getName())
+            return $this->builder
+                ->method($operation->getName())
                 ->makePublic()
                 ->addParam($methodParam)
                 ->addStmt($handleResponseStmt)
@@ -107,38 +112,51 @@ class ClientGenerator extends GeneratorAbstract
                 ->getNode();
         }
 
-        $mapperClassName = SchemaMapperNaming::getClassName($responseBody);
-        $this->addImport(
-            sprintf(
-                '%s%s\\%s',
-                $this->baseNamespace,
-                SchemaMapperGenerator::NAMESPACE_SUBPATH,
-                $mapperClassName
-            )
-        );
+        $responseVariable = $this->builder->var('response');
+        $stmts[]          = $this->builder->assign($responseVariable, $handleResponseStmt);
+        if ($responseBody->isComposite()) {
+            $mapperClassName = SchemaMapperNaming::getClassName($responseBody);
+            $this->addImport(
+                sprintf(
+                    '%s%s\\%s',
+                    $this->baseNamespace,
+                    SchemaMapperGenerator::NAMESPACE_SUBPATH,
+                    $mapperClassName
+                )
+            );
 
-        $getMethod = $this->builder->methodCall(
-            $this->builder->localPropertyFetch('container'),
-            'get',
-            [$this->builder->classConstFetch($mapperClassName, 'class')]
-        );
+            $getMethod = $this->builder->methodCall(
+                $this->builder->localPropertyFetch('container'),
+                'get',
+                [$this->builder->classConstFetch($mapperClassName, 'class')]
+            );
 
-        $mapMethod = $this->builder->methodCall($getMethod, 'toSchema', [$handleResponseStmt]);
-        $returnStmt = $this->builder->return($mapMethod);
+            $mapMethod = $this->builder->methodCall($getMethod, 'toSchema', [$responseVariable]);
+            $stmts[]   = $this->builder->return($mapMethod);
 
-        $this->addImport(
-            sprintf(
-                '%s%s\\%s',
-                $this->baseNamespace,
-                SchemaGenerator::NAMESPACE_SUBPATH,
-                $responseBody->getPhpClassName()
-            )
-        );
+            $this->addImport(
+                sprintf(
+                    '%s%s\\%s',
+                    $this->baseNamespace,
+                    SchemaGenerator::NAMESPACE_SUBPATH,
+                    $responseBody->getPhpClassName()
+                )
+            );
+        } else {
+            $this->addImport(CopiedNamespace::getImport($this->baseNamespace, ContentTypeSerializerInterface::class));
+            $literalValue = $this->builder->getArrayItem(
+                $responseVariable,
+                $this->builder->classConstFetch('ContentTypeSerializerInterface', 'LITERAL_VALUE_KEY')
+            );
 
-        return $this->builder->method($operation->getName())
+            $stmts[] = $this->builder->return($literalValue);
+        }
+
+        return $this->builder
+            ->method($operation->getName())
             ->makePublic()
             ->addParam($methodParam)
-            ->addStmt($returnStmt)
+            ->addStmts($stmts)
             ->setReturnType($responseBody->getPhpTypeHint(), $responseBody->isNullable())
             ->composeDocBlock([$methodParam], $responseBody->getPhpDocType(false))
             ->getNode();
@@ -198,7 +216,7 @@ class ClientGenerator extends GeneratorAbstract
             ->param('response')
             ->setType('ResponseInterface')
             ->getNode();
-        $response = $this->builder->var('response');
+        $response     = $this->builder->var('response');
 
         $handleResponseStatement = $this->builder->return(
             $this->builder->methodCall(