+++ src/Generator/MutatorAccessorClassGeneratorAbstract.php
@@ -9,7 +9,6 @@ use DoclerLabs\ApiClientGenerator\Entity\Field;
 use DoclerLabs\ApiClientGenerator\Input\Specification;
 use DoclerLabs\ApiClientGenerator\Naming\CopiedNamespace;
 use DoclerLabs\ApiClientGenerator\Naming\SchemaNaming;
-use DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType\Json\Json;
 use DoclerLabs\ApiClientGenerator\Output\Php\PhpFileCollection;
 use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\ClassMethod;
@@ -128,7 +127,6 @@ abstract class MutatorAccessorClassGeneratorAbstract extends GeneratorAbstract
         }
 
         $this
-            ->addImport(CopiedNamespace::getImport($this->baseNamespace, Json::class))
             ->addImport(RequestValidationException::class);
 
         $propertyVar       = $this->builder->var($root->getPhpVariableName());
@@ -150,7 +148,7 @@ abstract class MutatorAccessorClassGeneratorAbstract extends GeneratorAbstract
                 'Invalid %s value. Given: `%s` Allowed: %s',
                 $root->getName(),
                 $propertyVar,
-                $this->builder->staticCall('Json', 'encode', [$allowedConstFetch]),
+                $this->builder->funcCall('json_encode', [$allowedConstFetch]),
             ]
         );