+++ src/Output/Copy/Serializer/ContentType/ContentTypeSerializerInterface.php
@@ -7,7 +7,11 @@ use Psr\Http\Message\StreamInterface;
 
 interface ContentTypeSerializerInterface
 {
+    const LITERAL_VALUE_KEY = '__literalResponseValue';
+
     public function encode(SerializableInterface $body): string;
 
     public function decode(StreamInterface $body): array;
+
+    public function getMimeType(): string;
 }