+++ src/Output/Copy/Serializer/ContentType/Json/JsonException.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType\Json;
-
-use Exception;
-
-class JsonException extends Exception
-{
-}